requirements.txt
beautifulsoup4==4.14.2
EbookLib==0.20
lxml==6.0.2
PyPDF2==3.0.1
python-docx==1.2.0
six==1.17.0
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
